"sql data relaetionship"

+++ src/main/java/com/Unit2/OppHomeworkUnit2/model/MenuInput/Menu.java
@@ -25,7 +25,7 @@ public class Menu {
 
         String exit = null;
 
-        int id = 0;
+        Long id = 0L;
 
         System.out.println("List of available commands:\n- New Lead\n- Show Leads\n- Show Opportunities\n- Show Accounts" +
                 "\n- Lookup Lead ID\n- Lookup Opportunity ID\n- Lookup Account ID\n- Convert ID\n- Close Lost ID\n- Close Won ID\n- Help\n- Exit");
@@ -45,7 +45,7 @@ public class Menu {
                     command.contains("lookup lead") || command.contains("close lost") || command.contains("close won")) {
                 String[] splitCommand = command.split(" ");
                 try {
-                    id = Integer.parseInt(splitCommand[splitCommand.length - 1]);
+                    id = Long.parseLong(splitCommand[splitCommand.length - 1]);
                 }catch(NumberFormatException e){
                     command = "EXCEPTION";
                 }
@@ -57,7 +57,7 @@ public class Menu {
                 String[] splitCommand = command.split(" ");
                 command = splitCommand[0];
                 try {
-                    id = Integer.parseInt(splitCommand[splitCommand.length - 1]);
+                    id = Long.parseLong(splitCommand[splitCommand.length - 1]);
                 }catch(NumberFormatException e){
                     command = "EXCEPTION";
                     //En el caso que no se haya introducido un ID correcto saltara una excepcion
